requirements.txt
daphne
Django==5.0
sqlparse==0.4.4
tzdata==2024.1
psycopg2-binary==2.9.6
djangorestframework==3.15.2
django-cors-headers
celery[redis]
drf_yasg==1.21.7
databases[postgresql]
python-utils
django-rest-swagger
drf-yasg
elasticsearch_dsl==8.14.0
python-dotenv==1.0.1
whitenoise